temp/ai/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
torch>=1.9.0
web3>=6.0.0
pymongo>=4.0.0
websockets>=10.0
python-dotenv>=0.19.0
tenacity>=8.0.0
joblib>=1.0.0
aiohttp>=3.8.0
asyncio>=3.4.3
matplotlib>=3.4.0
seaborn>=0.11.0
